requirements.txt
# Libraries used in visualize.py
pygame==2.1.3
python-chess==1.999
torch==2.0.1
numpy==1.25.0
requests==2.31.0